requirements.txt
Django>=4.2
djangorestframework
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth
requests
django-cors-headers